requirements.txt
requests==2.32.3
tqdm==4.66.6
gdown==5.2.0
pillow==11.0.0
matplotlib==3.9.2
langdetect==1.0.9
pandas==2.2.3
torch==2.5.1
torchvision==0.20.1
transformers==4.47.0
gradio==5.9.0
ruff==0.8.3
ruff-lsp==0.0.59
streamlit==1.41.1
